+++ src/main/java/wooteco/subway/admin/service/PathService.java
@@ -1,14 +1,15 @@
 package wooteco.subway.admin.service;
 
-import org.jgrapht.GraphPath;
 import org.springframework.stereotype.Service;
-import wooteco.subway.admin.domain.*;
+import wooteco.subway.admin.domain.EdgeType;
+import wooteco.subway.admin.domain.Graph;
+import wooteco.subway.admin.domain.Line;
+import wooteco.subway.admin.domain.Station;
 import wooteco.subway.admin.dto.SearchPathResponse;
 import wooteco.subway.admin.repository.LineRepository;
 import wooteco.subway.admin.repository.StationRepository;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import static wooteco.subway.admin.domain.EdgeType.DISTANCE;
 import static wooteco.subway.admin.domain.EdgeType.DURATION;
@@ -32,34 +33,12 @@ public class PathService {
         Station targetStation = findStationByName(targetStationName, stations);
 
         Graph graph = Graph.of(lines, stations, EdgeType.of(type));
-        GraphPath<Station, Edge> shortestPath = graph.findShortestPath(startStation, targetStation);
 
-        validatePath(shortestPath);
-
-        int durationSum = getEdgeValueSum(shortestPath, DURATION);
-        int distanceSum = getEdgeValueSum(shortestPath, DISTANCE);
-        List<String> stationNames = getVertexName(shortestPath);
+        int durationSum = graph.getEdgeValueSum(startStation, targetStation, DURATION);
+        int distanceSum = graph.getEdgeValueSum(startStation, targetStation, DISTANCE);
+        List<String> stationNames = graph.getVertexName(startStation, targetStation);
 
         return new SearchPathResponse(durationSum, distanceSum, stationNames);
-    }
-
-    private int getEdgeValueSum(GraphPath<Station, Edge> shortestPath, EdgeType edgeType) {
-        return shortestPath.getEdgeList().stream()
-                .mapToInt(edge -> edgeType.getEdgeValue(edge.toLineStation()))
-                .sum();
-    }
-
-    private List<String> getVertexName(GraphPath<Station, Edge> shortestPath) {
-        return shortestPath.getVertexList().stream()
-                .map(edge -> edge.getName())
-                .collect(Collectors.toList());
-    }
-
-    private Station findStationByName(String stationName, List<Station> stations) {
-        return stations.stream()
-                .filter(station -> stationName.equals(station.getName()))
-                .findAny()
-                .orElseThrow(() -> new IllegalArgumentException("역이 존재하지 않습니다."));
     }
 
     private void validateStationSame(String startStationName, String targetStationName) {
@@ -68,9 +47,10 @@ public class PathService {
         }
     }
 
-    private void validatePath(GraphPath<Station, Edge> path) {
-        if (path == null) {
-            throw new IllegalArgumentException("두 역이 연결되어있지 않습니다.");
-        }
+    private Station findStationByName(String stationName, List<Station> stations) {
+        return stations.stream()
+                .filter(station -> stationName.equals(station.getName()))
+                .findAny()
+                .orElseThrow(() -> new IllegalArgumentException("역이 존재하지 않습니다."));
     }
 }
